+++ pkg/version/version.go
@@ -1,0 +1,8 @@
+package version
+
+var (
+	GitSource   string
+	GitTag      string
+	GetBranch   string
+	GoBuildTime string
+)
